Fixed view for a date in ForecastActivity. Changed implementation for CurrentWeather in MainActivity - for now GPSService send intent for WeatherService and WeatherService send result for MainActivity using EventBus.

+++ app/src/main/java/pl/pb/r/kcksm/model/weather/Day.java
@@ -13,7 +13,7 @@ import java.util.List;
 public class Day {
     @SerializedName("dt")
     @Expose
-    public Integer dt;
+    public Long dt;
     @SerializedName("temp")
     @Expose
     public Temp temp;
